Trigger unit tests on change of kubo-deployment
The intent here is we do not need to trigger the unit tests because they
all live in kubo-release. However- that's not entirely true because the
task itself (run-unit-tests.yml) is in git-kubo-deployment so we should
still run it.

Without that additional trigger we'll still run the pipeline for
kubo-deployment changes but we won't exercise the task if it changed.

[#137357741]