Remove outdated docs and refactor terraform examples
This change removes a large majority of the `docs/`. It still contains a
`README.md` that refers users to the official documentation at
https://docs-cfcr.cfapps.io/.

This change also moves the terraform templates to a `docs/terraform`
directory as they are still referenced by the official documentation.

This change also removes the `setup_helpers.sh` that previously existed
for AWS and uses the "file" provisioner to provide these files on the
bastion. This change brings the AWS terraform process closer to the GCP
process.

[#153669766]

Signed-off-by: Brenda Chan <brchan@pivotal.io>